+++ Logos/src/utopia/logos/model/stored/url/StoredLink.scala
@@ -13,12 +13,12 @@ object StoredLink extends StoredFromModelFactory[LinkData, StoredLink]
 	
 	private lazy val questionMarkRegex = Regex.escape('?')
 	private lazy val pathCharacterRegex = 
-		(Regex.letterOrDigit || Regex.anyOf("-._~:/#[]@!$&'()*+,;%=")).withinParenthesis
-	private lazy val urlCharacterRegex = (pathCharacterRegex || questionMarkRegex).withinParenthesis
+		(Regex.letterOrDigit || Regex.anyOf("-._~:/#[]@!$&'()*+,;%=")).withinParentheses
+	private lazy val urlCharacterRegex = (pathCharacterRegex || questionMarkRegex).withinParentheses
 	/**
 	  * A regular expression that matches to the parameters -part of a link
 	  */
-	lazy val paramPartRegex = (questionMarkRegex + urlCharacterRegex.oneOrMoreTimes).withinParenthesis
+	lazy val paramPartRegex = (questionMarkRegex + urlCharacterRegex.oneOrMoreTimes).withinParentheses
 	/**
 	  * A regular expression that matches to links
 	  */